Bug 30707: Move container's closing div tag into template from intranet-bottom.inc
For templates which use the intranet-bottom include file, the markup in
that template lacks the main container's closing </div>, expecting it to
be closed in intranet-bottom.inc.

This patch moves that closing </div> into each template. This is
intended to make it clearer to anyone trying to maintain the correct
structure of the page and hopefully to make it possible to automatically
tidy template files.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit ec0aeda9c7d5ff9b6f9ae9c4488166017d806ae0)