Bug 10190: Rework logic of overdue_notices
The original logic for digesting overdues in the overdue notices script
was flawed given the additional itemtype handling introduced in this
patchset and the way in which it handles effective fallbacks for rules.

This patch tidies the file and reworks the logic so that we fetch
overdues by patron category and branch and then iterate through those
comparing to the correct effective rules and combing them into digest
groups based on notice + trigger level.

(cherry picked from commit 8eb165fbd80ff6c537a07091ab96c74a0a47a3bf)