Bug 10190: Modal form improvements
We now fetch all rules on modal open and the effective rule on context
selection.

This allows us to compare the specific values of the rule we're
adding/editing to the fallback values present in the all rules set.

We use the fallback rule values as placeholders in the input boxes to
highlight to the end user what value will be used should they choose to
wipe the input.  We also allow wiping inputs and pass 'null' to the API
appropriately.

Finally, we set a minimum value for delay when adding a new rule to
highlight to the user that the new trigger will be added after existing
triggers.

(cherry picked from commit 7904425c9bbe892f8370c5f58a1e3049db12afb8)