Bug 10190: Rename 'All *' to 'Default rule for *'
Try to clarify that the 'All' rules are actually fallback/default rules
for when there's not a more specific rule for the context.

We need to expose these clearly for each 'Library' view too which we're
not currently doing.

Sponsored-by: Glasgow Colleges Library Group <https://library.cityofglasgowcollege.ac.uk>
Signed-off-by: George Harkins <George.Harkins@cityofglasgowcollege.ac.uk>
(cherry picked from commit 46050d3a46d9cef256d8523df639fd14e55b3237)