Bug 38924: Add quota picker to checkout confirmation
This patch adds the quota selector to the checkout confirmation for if
multiple eligable quotas are available

Bug 38924: Return all eligable quotas so that the staff user can pick

This patch adds the logic to return multiple eligable quotas if there
are more than 1, allowing the staff user to pick which to assign an item
to.

Bug 38924: Add searcher method for currently active quota for patron

Bug 38924: amend circ logic to handle multiple eligable quotas

This amends the circulations logic to handle the fact that we can return
multiple eligable quotas now. This ensures that the same quotas as was
used for the checkout will be used for the renewal (both check and
assignment) and that we use the selected quota from our confirmation
dropdown for checkout of new item

Bug 38924: add better name display for quota select at checkout