Bug 39497: update 'Lists' page tabs attributes
On the Koha OPAC, the 'Lists' page contains tabs which aren't clearly marked as such (no relevant attribute set on the element).

This means that screen readers will not recognise the tabs as such, and screen reader users will not be made aware that these are available.

Since the 'Advanced Search' page contains tabs that are correctly identified and described by screen readers, update the HTML atributes for the tabs on the 'Lists' page so that it matches theirs.

Note: this is a quick and simple fix aiming to meet accessibility requirements. Another approach could be to implement wrappers (as in advsearch.tt), but it seems a bigger refactor of opac-shelves.tt may be needed to allow that.

Test plan:

0) Before testing, ensure that the relevant patron has a few public and private lists.
1) Log in on the Koha OPAC
3) Navigate to Home > koha > Lists
2) Notice that, when the page first load, the 'Your Lists' tab panel is displayed (functionality is preserved)
3) Select 'Public lists', and notice that the tab panel now matches that (again, functionality is preserved)
4) Either inspect the page's HTML or run a Lighthouse test, and check that all aria attribute required by screen readers are present

(aria attributes assumed to be required are aria-controls, aria-selected, and aria-labelledby. It is also assumed that 'role' attributes are needed)