Bug 10190: fix: allow more than 9 triggers
note: --no-verify was necessary to override the pre-commit hook
blocking submission due to the following:

Constant subroutine Koha::CirculationRules::GUESSED_ITEMTYPES_KEY
redefined at /usr/lib/x86_64-linux-gnu/perl-base/constant.pm line 171.

To my understanding, this is not linked to the changes in this commit,
and using --no-verify is likely fine in this case, though I'd be
happy to be corrected on that!