Bug 39547: (follow-up) wrapped required labels in conditional-ifs
This patch wraps the required_label lines in conditional-ifs, which
means if they are not applicable, they will never reach the client. This
patch has been written following customer feedback, in which it was
noted that the presence of these elements in the DOM would trigger a
screenreader to read them out, as if they were on the page.