Bug 37893: (QA follow-up): Add tooltip for account Staff userid
The userid field for accounts was missing a tooltip; We add one here
that clarifies that the userid must match an existing Koha user with
appropriate circulation permissions for the sip client.

NOTE: A future follow-up should introduce a patron autocomplete or
select type input allowing administrators to just search users to add
for this field, but such a component doesn't exist in the Vue
architecture yet.

Signed-off-by: Andrew Fuerste Henry <andrew@bywatersolutions.com>
Signed-off-by: Lisette Scheer <lisette@bywatersolutions.com>
Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>