Bug 41178: (VueJS framework): Add i18n.deleteConfirmationBody
Otherwise, we're unable to override the contents of the delete dialog

This and the previous patch allow fetching related accounts on the fly
when attempting to delete an institutions. This allows to provide a
message stating how many (or which) accounts will be deleted if the
selected institution is deleted.

The way its implemented applies for both deleting the resource from the
list or from the 'show' page.

If the institution does not have any associated accounts, the delete
dialog's behavior is the same as it was before.